+++ src/domains/peers/mod.rs
@@ -1,5 +1,5 @@
 pub mod connected_peer_info;
 pub(crate) mod connected_types;
 pub(crate) mod identifier;
-pub(crate) mod kind;
+
 pub(crate) mod peer;
