+++ src/presentation/cluster_in/outbound/stream.rs
@@ -4,8 +4,8 @@ use crate::domains::cluster_actors::commands::ClusterCommand;
 use crate::domains::cluster_actors::replication::ReplicationInfo;
 use crate::domains::peers::connected_peer_info::ConnectedPeerInfo;
 use crate::domains::peers::identifier::PeerIdentifier;
-use crate::domains::peers::kind::PeerKind;
 use crate::domains::peers::peer::Peer;
+use crate::domains::peers::peer::PeerKind;
 use crate::presentation::cluster_in::connection_manager::ClusterConnectionManager;
 use crate::services::interface::TRead;
 use crate::services::interface::TWrite;
